import-export.cpp: fixes a signed/unsigned comparaison issue

+++ src/import-export.cpp
@@ -154,7 +154,7 @@ ppk_error ImportExport::dumpModuleToFile(const char* module_id, const char* file
 	if(res!=PPK_OK)
 		return res;
 
-	for(int i=0; i<nbEntries; i++)
+	for(size_t i=0; i<nbEntries; i++)
 	{
 		ppk_data* edata;
